server/requirements.txt
-i https://pypi.org/simple
alembic==1.14.1; python_version >= '3.8'
blinker==1.8.2; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
flask-cors==5.0.0
flask-migrate==4.1.0; python_version >= '3.6'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
importlib-metadata==8.5.0; python_version >= '3.8'
importlib-resources==6.4.5; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.10; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sqlalchemy==2.0.41; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
werkzeug==3.0.6; python_version >= '3.8'
zipp==3.20.2; python_version >= '3.8'
